fix port type (smallint -> integer)

+++ migrations/188_workspaces__port__add.sql
@@ -1,1 +1,1 @@
-ALTER TABLE workspaces ADD COLUMN IF NOT EXISTS port smallint;
+ALTER TABLE workspaces ADD COLUMN IF NOT EXISTS port integer;
